Fix the two native-binding bugs the Unix audit found
Both were verified against the real headers on the GCC compile farm rather than inferred.

FreeBSD's `struct utmpx` is 280 bytes, not 276: ut_tv aligns it to 8 and structLayout
never infers tail padding. Every field offset was already right, so this was latent, in
the same way the Windows tail-padding errors were. Its JNA twin nests a Timeval and so
gets the size from JNA's own alignment.

AIX `thread_self` returns tid_t, which follows long -- measured at 8 bytes under -maix64
and 4 under -maix32. OSHI runs 64-bit there, and both twins declared it 4, reading only
the low half of the returned register. Declare it 64-bit and narrow at the API boundary,
since OSThread.getThreadId() is int and cannot widen.

Nest ut_tv in the macOS, Linux, Solaris and FreeBSD utmpx layouts rather than flattening
it to ut_tv_sec/ut_tv_usec, and give Linux's two spare arrays the names glibc declares
(__f_spare, __glibc_reserved). A member the header does not declare cannot be checked, so
those six offsets were being skipped on every platform -- on the very struct that turned
out to have a real bug on FreeBSD. Offsets are unchanged; macOS and Linux now check 164
and 43 field offsets with none unverifiable, against 163 and 40 with six skipped.

+++ oshi-core/src/main/java/oshi/jna/platform/unix/AixLibc.java
@@ -19,5 +19,5 @@ public interface AixLibc extends CLibrary {
      *
      * @return the caller's kernel thread ID.
      */
-    int thread_self();
+    long thread_self();
 }
