+++ oshi-core/src/main/java/oshi/software/os/unix/aix/AixOperatingSystemJNA.java
@@ -59,7 +59,8 @@ public final class AixOperatingSystemJNA extends AixOperatingSystem {
 
     @Override
     public int getThreadId() {
-        return AixLibc.INSTANCE.thread_self();
+        // OSThread.getThreadId() is int; AIX tid_t is 64-bit, so narrow at the API boundary
+        return (int) AixLibc.INSTANCE.thread_self();
     }
 
     @Override
